Expose authenticate
